Fixed trap damage, collisions, added interaction with ending
As well as updating the prefab on some traps, redesigning some trap layouts in the rooms and also made a start on optimising the map from the screen tearing where 2 wall blocks overlap and fight for the same world position. Added 2 scripts that play the animations for the chest and the trapdoor at the end (maybe add a short cinematic / score total on screen here?) and rebuilt the project with "Level 1" and "Level 2". Trapdoor at the end of level 1 changes the scene to 2.